src/requirements.txt
azure_core
azure-ai-documentintelligence
streamlit
azure-storage-blob
python-dotenv